@@ -1,6 +1,5 @@
 {
   "name": "@redux-eggs-example/next",
-  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "cross-env PORT=3010 next",
@@ -9,21 +8,7 @@
     "analyze": "cross-env ANALYZE=true yarn build",
     "lint": "next lint",
     "type-check": "tsc --pretty --noEmit --incremental false",
-    "test": "jest",
-    "test-all": "yarn lint && yarn type-check && yarn test"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "test": "yarn lint && yarn type-check"
   },
   "dependencies": {
     "@redux-eggs/core": "workspace:*",
@@ -31,44 +16,29 @@
     "@redux-eggs/react": "workspace:*",
     "@redux-eggs/redux": "workspace:*",
     "@redux-eggs/saga-extension": "workspace:*",
-    "@redux-saga/is": "1.1.2",
-    "@redux-saga/symbols": "1.1.2",
-    "clsx": "1.1.1",
     "cross-env": "7.0.3",
-    "dayjs": "1.10.7",
-    "immer": "9.0.12",
-    "mitt": "3.0.0",
-    "nanoid": "3.3.1",
-    "next": "12.1.0",
-    "nprogress": "0.2.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-intersection-observer": "8.33.1",
-    "react-redux": "7.2.6",
-    "redux": "4.1.2",
+    "immer": "9.0.15",
+    "next": "12.3.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-intersection-observer": "9.3.5",
+    "react-redux": "8.0.2",
+    "redux": "4.2.0",
     "redux-devtools-extension": "2.13.9",
     "redux-logger": "3.0.6",
-    "redux-saga": "1.1.3",
-    "reselect": "4.1.5"
+    "redux-saga": "1.2.3",
+    "reselect": "4.1.6"
   },
   "devDependencies": {
-    "@jest/types": "27.5.1",
-    "@next/bundle-analyzer": "12.1.0",
-    "@types/jest": "27.4.0",
-    "@types/node": "17.0.18",
-    "@types/nprogress": "0.2.0",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
-    "@types/react-redux": "7.1.22",
+    "@next/bundle-analyzer": "12.2.2",
+    "@types/node": "20.1.0",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
     "@types/redux-logger": "3.0.9",
-    "eslint": "8.9.0",
-    "eslint-config-next": "12.1.0",
-    "jest": "27.5.1",
-    "redux-saga-test-plan": "4.0.4",
-    "sass": "1.49.8",
-    "styled-jsx": "5.0.0",
-    "ts-jest": "27.1.3",
-    "tslib": "2.3.1",
-    "typescript": "4.5.5"
+    "eslint": "8.19.0",
+    "eslint-config-next": "12.2.2",
+    "sass": "1.62.1",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4"
   }
 }
