@@ -1,46 +1,42 @@
 {
   "name": "@redux-eggs-example/react",
-  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview",
     "lint": "eslint src",
-    "type-check": "tsc --pretty --noEmit --incremental false",
+    "type-check": "tsc --pretty --noEmit",
     "test": "yarn lint && yarn type-check"
   },
   "dependencies": {
     "@redux-eggs/core": "workspace:*",
     "@redux-eggs/react": "workspace:*",
     "@redux-eggs/redux": "workspace:*",
-    "history": "5.2.0",
-    "immer": "9.0.7",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-intersection-observer": "8.32.0",
-    "react-redux": "7.2.6",
-    "react-router": "6.2.1",
-    "react-router-dom": "6.2.1",
-    "redux": "4.1.2",
+    "immer": "9.0.15",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-intersection-observer": "9.3.5",
+    "react-redux": "8.0.2",
+    "react-router": "6.3.0",
+    "react-router-dom": "6.3.0",
+    "redux": "4.2.0",
     "redux-devtools-extension": "2.13.9",
     "redux-logger": "3.0.6",
     "redux-thunk": "2.4.1",
-    "reselect": "4.1.5"
+    "reselect": "4.1.6"
   },
   "devDependencies": {
-    "@types/node": "17.0.5",
-    "@types/react": "17.0.37",
-    "@types/react-dom": "17.0.11",
-    "@types/react-redux": "7.1.20",
-    "@types/redux-logger": "3.0.8",
-    "@vitejs/plugin-react": "1.1.3",
-    "eslint": "7.32.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.28.0",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "tslib": "2.3.1",
-    "typescript": "4.5.2",
-    "vite": "2.7.7"
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
+    "@types/redux-logger": "3.0.9",
+    "@vitejs/plugin-react": "4.0.0",
+    "eslint": "8.19.0",
+    "eslint-plugin-jsx-a11y": "6.6.0",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4",
+    "vite": "4.3.5"
   }
 }
