@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "main": "dist/index.js",
-  "types": "types/index.d.ts",
+  "types": "dist/index.d.ts",
   "files": [
     "dist"
   ],
@@ -16,23 +16,23 @@
   "dependencies": {
     "brotli-size": "4.0.0",
     "cli-table": "0.3.11",
-    "colorette": "2.0.16",
-    "filesize": "8.0.7",
-    "pacote": "13.0.2",
-    "terser": "5.10.0"
+    "colorette": "2.0.19",
+    "filesize": "9.0.11",
+    "pacote": "13.6.2",
+    "terser": "5.15.0"
   },
   "devDependencies": {
-    "@rollup/plugin-eslint": "8.0.1",
-    "@types/cli-table": "0.3.0",
-    "@types/node": "17.0.18",
-    "@types/pacote": "11.1.3",
-    "eslint": "8.9.0",
+    "@rollup/plugin-eslint": "8.0.2",
+    "@types/cli-table": "0.3.1",
+    "@types/node": "20.1.0",
+    "@types/pacote": "11.1.5",
+    "eslint": "8.24.0",
     "rimraf": "3.0.2",
-    "rollup": "2.67.3",
+    "rollup": "2.79.1",
     "rollup-plugin-peer-deps-external": "2.2.4",
-    "rollup-plugin-ts": "2.0.5",
-    "tslib": "2.3.1",
-    "typescript": "4.5.5"
+    "rollup-plugin-ts": "3.0.2",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4"
   },
   "peerDependencies": {
     "rollup": "^2.57.0"
