@@ -15,39 +15,42 @@
     "build": "lerna run --stream --concurrency 1 --sort build",
     "build:size": "lerna run --stream --concurrency 1 --sort build:size",
     "clean": "lerna run --parallel clean",
+    "clean:dist": "lerna run --parallel clean:dist",
+    "clean:coverage": "lerna run --parallel clean:coverage",
+    "clean:file-size-cache": "lerna run --parallel clean:file-size-cache",
     "lint": "lerna run --parallel lint",
     "type-check": "lerna run --parallel type-check",
     "test": "lerna run --parallel test",
     "test:coverage": "lerna run --parallel test:coverage",
     "prepublish": "lerna run prepublish",
-    "postinstall": "husky install",
-    "pretty-quick": "pretty-quick"
+    "postinstall": "husky install"
   },
   "engines": {
     "node": ">=16",
     "yarn": ">=3"
   },
-  "packageManager": "yarn@3.2.0",
+  "packageManager": "yarn@3.5.1",
   "workspaces": [
     "examples/*",
     "internal/*",
     "packages/*"
   ],
   "devDependencies": {
-    "@commitlint/cli": "16.2.1",
-    "@commitlint/config-conventional": "16.2.1",
-    "@typescript-eslint/eslint-plugin": "5.12.0",
-    "@typescript-eslint/parser": "5.12.0",
-    "eslint": "8.9.0",
-    "eslint-config-prettier": "8.4.0",
-    "eslint-import-resolver-typescript": "2.5.0",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-simple-import-sort": "7.0.0",
-    "eslint-plugin-unicorn": "41.0.0",
-    "husky": "7.0.4",
-    "lerna": "4.0.0",
-    "prettier": "2.5.1",
+    "@commitlint/cli": "17.1.2",
+    "@commitlint/config-conventional": "17.1.0",
+    "@typescript-eslint/eslint-plugin": "5.38.1",
+    "@typescript-eslint/parser": "5.38.1",
+    "eslint": "8.24.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-import-resolver-typescript": "3.5.1",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-simple-import-sort": "8.0.0",
+    "eslint-plugin-unicorn": "43.0.2",
+    "husky": "8.0.1",
+    "lerna": "5.5.4",
+    "prettier": "2.7.1",
     "pretty-quick": "3.1.3",
-    "typescript": "4.5.5"
+    "typescript": "4.6.4"
   }
 }
