@@ -24,7 +24,10 @@
   "scripts": {
     "build": "rollup -c",
     "build:size": "cross-env FILESIZE=true yarn run prepublish",
-    "clean": "rimraf coverage dist file-size-cache",
+    "clean": "rimraf dist coverage file-size-cache",
+    "clean:dist": "rimraf dist",
+    "clean:coverage": "rimraf coverage",
+    "clean:file-size-cache": "rimraf file-size-cache",
     "lint": "eslint src",
     "type-check": "tsc --pretty --noEmit",
     "test": "jest",
@@ -32,21 +35,21 @@
     "prepublish": "yarn run clean && yarn run build"
   },
   "devDependencies": {
-    "@jest/types": "27.5.1",
+    "@jest/types": "29.1.0",
     "@redux-eggs-internal/rollup-plugin-filesize": "1.0.0",
-    "@rollup/plugin-eslint": "8.0.1",
-    "@types/jest": "27.4.0",
+    "@rollup/plugin-eslint": "8.0.2",
+    "@types/jest": "28.1.5",
     "cross-env": "7.0.3",
-    "jest": "27.5.1",
-    "redux": "4.1.2",
+    "jest": "29.5.0",
+    "jest-environment-jsdom": "29.5.0",
     "rimraf": "3.0.2",
-    "rollup": "2.67.3",
+    "rollup": "2.77.0",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-terser": "7.0.2",
-    "rollup-plugin-ts": "2.0.5",
-    "ts-jest": "27.1.3",
-    "tslib": "2.3.1",
-    "typescript": "4.5.5"
+    "rollup-plugin-ts": "3.0.2",
+    "ts-jest": "29.0.2",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4"
   },
   "gitHead": "5ead04c987c52df518ab3558168e1595df3dde19"
 }
