@@ -24,7 +24,10 @@
   "scripts": {
     "build": "rollup -c",
     "build:size": "cross-env FILESIZE=true yarn run prepublish",
-    "clean": "rimraf coverage dist file-size-cache",
+    "clean": "rimraf dist coverage file-size-cache",
+    "clean:dist": "rimraf dist",
+    "clean:coverage": "rimraf coverage",
+    "clean:file-size-cache": "rimraf file-size-cache",
     "lint": "eslint src",
     "type-check": "tsc --pretty --noEmit",
     "test": "jest",
@@ -32,43 +35,44 @@
     "prepublish": "yarn run clean && yarn run build"
   },
   "devDependencies": {
-    "@jest/types": "27.5.1",
+    "@jest/types": "29.1.0",
     "@redux-eggs-internal/rollup-plugin-filesize": "1.0.0",
     "@redux-eggs/core": "2.2.0",
     "@redux-eggs/redux": "2.2.0",
-    "@rollup/plugin-eslint": "8.0.1",
-    "@testing-library/react": "12.1.3",
-    "@types/jest": "27.4.0",
-    "@types/node": "17.0.18",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
-    "@types/react-redux": "7.1.22",
-    "@types/react-test-renderer": "17.0.1",
+    "@rollup/plugin-eslint": "8.0.2",
+    "@testing-library/react": "13.4.0",
+    "@types/jest": "29.0.3",
+    "@types/node": "20.1.0",
+    "@types/react": "18.0.21",
+    "@types/react-dom": "18.0.6",
+    "@types/react-test-renderer": "18.0.0",
     "cross-env": "7.0.3",
-    "eslint": "8.9.0",
-    "eslint-plugin-react": "7.28.0",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "jest": "27.5.1",
-    "next": "12.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.6",
-    "react-test-renderer": "17.0.2",
-    "redux": "4.1.2",
+    "eslint": "8.24.0",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "jest": "29.5.0",
+    "jest-environment-jsdom": "29.5.0",
+    "next": "12.3.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-redux": "8.0.4",
+    "react-test-renderer": "18.2.0",
+    "redux": "4.2.0",
     "rimraf": "3.0.2",
-    "rollup": "2.67.3",
+    "rollup": "2.79.1",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-terser": "7.0.2",
-    "rollup-plugin-ts": "2.0.5",
-    "ts-jest": "27.1.3",
-    "tslib": "2.3.1",
-    "typescript": "4.5.5"
+    "rollup-plugin-ts": "3.0.2",
+    "ts-jest": "29.0.2",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4"
   },
   "peerDependencies": {
     "@redux-eggs/core": "^2.0.0",
-    "next": ">=11.1.0",
-    "react": "^16.8.3 || ^17.0.0",
-    "react-redux": "^7.0.0"
+    "next": "^11.1.0 || ^12.0.0",
+    "react": "^16.8.3 || ^17.0.0 || ^18.0.0",
+    "react-redux": "^7.0.0 || ^8.0.0",
+    "redux": "^4.1.2"
   },
   "gitHead": "5ead04c987c52df518ab3558168e1595df3dde19"
 }
