@@ -11,9 +11,9 @@
     "url": "https://github.com/fostyfost/redux-eggs/issues"
   },
   "homepage": "https://github.com/fostyfost/redux-eggs",
-  "main": "dist/index.js",
-  "module": "dist/index.es.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "dist/es/index.js",
+  "types": "./dist/cjs/index.d.ts",
   "files": [
     "dist"
   ],
@@ -23,7 +23,10 @@
   "scripts": {
     "build": "rollup -c",
     "build:size": "cross-env FILESIZE=true yarn run prepublish",
-    "clean": "rimraf coverage dist file-size-cache",
+    "clean": "rimraf dist coverage file-size-cache",
+    "clean:dist": "rimraf dist",
+    "clean:coverage": "rimraf coverage",
+    "clean:file-size-cache": "rimraf file-size-cache",
     "lint": "eslint src",
     "type-check": "tsc --pretty --noEmit",
     "test": "jest",
@@ -31,41 +34,43 @@
     "prepublish": "yarn run clean && yarn run build"
   },
   "devDependencies": {
-    "@jest/types": "27.5.1",
+    "@jest/types": "29.1.0",
     "@redux-eggs-internal/rollup-plugin-filesize": "1.0.0",
     "@redux-eggs/core": "2.2.0",
-    "@rollup/plugin-eslint": "8.0.1",
-    "@testing-library/dom": "8.11.3",
-    "@testing-library/jest-dom": "5.16.2",
-    "@testing-library/react": "12.1.3",
-    "@testing-library/user-event": "13.5.0",
-    "@types/jest": "27.4.0",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
-    "@types/react-redux": "7.1.22",
-    "@types/testing-library__jest-dom": "5.14.2",
+    "@redux-eggs/redux": "2.2.0",
+    "@redux-eggs/redux-toolkit": "2.2.0",
+    "@rollup/plugin-eslint": "8.0.2",
+    "@testing-library/dom": "8.18.1",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
+    "@types/jest": "29.0.3",
+    "@types/react": "18.0.21",
+    "@types/react-dom": "18.0.6",
+    "@types/testing-library__jest-dom": "5.14.5",
     "cross-env": "7.0.3",
-    "eslint": "8.9.0",
-    "eslint-plugin-react": "7.28.0",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "jest": "27.5.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.6",
-    "redux": "4.1.2",
+    "eslint": "8.24.0",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "jest": "29.5.0",
+    "jest-environment-jsdom": "29.5.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-redux": "8.0.4",
+    "redux": "4.2.0",
     "rimraf": "3.0.2",
-    "rollup": "2.67.3",
+    "rollup": "2.79.1",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-terser": "7.0.2",
-    "rollup-plugin-ts": "2.0.5",
-    "ts-jest": "27.1.3",
-    "tslib": "2.3.1",
-    "typescript": "4.5.5"
+    "rollup-plugin-ts": "3.0.2",
+    "ts-jest": "29.0.2",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4"
   },
   "peerDependencies": {
     "@redux-eggs/core": "^2.0.0",
-    "react": "^16.8.3 || ^17.0.0",
-    "react-redux": "^7.0.0"
+    "react": "^16.8.3 || ^17.0.0 || ^18.0.0",
+    "react-redux": "^7.0.0 || ^8.0.0"
   },
   "gitHead": "5ead04c987c52df518ab3558168e1595df3dde19"
 }
