@@ -1,12 +1,13 @@
 import { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 
 import { App } from './app'
 import { store } from './store'
 
-ReactDOM.render(
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <Provider store={store}>
       <BrowserRouter>
@@ -14,5 +15,4 @@ ReactDOM.render(
       </BrowserRouter>
     </Provider>
   </StrictMode>,
-  document.getElementById('root'),
 )
