@@ -1,30 +1,44 @@
-// TODO eslint-plugin-import, unicorn
-
 module.exports = {
   root: true,
   parser: '@typescript-eslint/parser',
   extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended', 'prettier'],
-  plugins: ['@typescript-eslint', 'simple-import-sort', 'unicorn'],
+  plugins: ['@typescript-eslint', 'simple-import-sort', 'unicorn', 'prettier'],
   rules: {
     'unicorn/filename-case': 'error',
-    'no-use-before-define': 'off',
-    '@typescript-eslint/no-use-before-define': 'error',
-    '@typescript-eslint/consistent-type-imports': 'error',
-    '@typescript-eslint/no-unused-vars': 'error',
     'no-console': 'off',
     'linebreak-style': ['error', 'unix'],
     'no-multiple-empty-lines': ['error', { max: 1, maxEOF: 0, maxBOF: 0 }],
     strict: ['error', 'global'],
     curly: 'warn',
-    quotes: ['error', 'single'],
-    'simple-import-sort/imports': 'error',
-    'simple-import-sort/exports': 'error',
+    quotes: ['error', 'single', { avoidEscape: true }],
     'sort-imports': 'off',
     'import/order': 'off',
+    'simple-import-sort/imports': 'error',
+    'simple-import-sort/exports': 'error',
+    'no-use-before-define': 'off',
+    '@typescript-eslint/no-use-before-define': [
+      'error',
+      {
+        enums: false,
+        typedefs: false,
+        ignoreTypeReferences: true,
+      },
+    ],
+    '@typescript-eslint/consistent-type-imports': 'error',
+    '@typescript-eslint/no-unused-vars': 'error',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/ban-ts-comment': 'off',
+    '@typescript-eslint/ban-types': [
+      'error',
+      {
+        types: {
+          '{}': false,
+        },
+      },
+    ],
+    'prettier/prettier': 'error',
   },
   settings: {
     'import/resolver': {
