@@ -1,9 +1,9 @@
-import type { FC } from 'react'
+import type { FC, PropsWithChildren } from 'react'
 
 import { Navigation } from '@/components/layout/navigation'
 import { Xkcd } from '@/components/xkcd'
 
-const Layout: FC = ({ children }) => {
+const Layout: FC<PropsWithChildren> = ({ children }) => {
   return (
     <>
       <p>
