@@ -1,4 +1,4 @@
-import { withEggs } from '@redux-eggs/react'
+import { getInjector } from '@redux-eggs/react'
 import NextLink from 'next/link'
 import type { FC } from 'react'
 import { useInView } from 'react-intersection-observer'
@@ -26,16 +26,24 @@ const LoadedPosts: FC<Props> = ({ rootUrl }) => {
   )
 }
 
-const postEgg = getPostsEgg({ afterAdd: store => store.dispatch(PostsPublicAction.loadPosts()) })
-
-export const Posts: FC<Props> = withEggs<Props>([postEgg])(function Posts({ rootUrl }) {
+export const Posts: FC<Props> = ({ rootUrl }) => {
   const isLoading = useSelector(isPostsLoading)
 
   return isLoading ? <div>Posts are loading ...</div> : <LoadedPosts rootUrl={rootUrl} />
-})
+}
+
+const Injector = getInjector([getPostsEgg({ afterAdd: store => store.dispatch(PostsPublicAction.loadPosts()) })])
 
 export const DynamicPosts: FC<Props> = ({ rootUrl }) => {
   const [ref, inView] = useInView()
 
-  return <div ref={ref}>{inView ? <Posts rootUrl={rootUrl} /> : null}</div>
+  return (
+    <div ref={ref}>
+      {inView ? (
+        <Injector.Wrapper>
+          <Posts rootUrl={rootUrl} />
+        </Injector.Wrapper>
+      ) : null}
+    </div>
+  )
 }
