@@ -25,7 +25,7 @@ function* loadXkcdInfoWorker() {
 
   try {
     // Alternative api https://xkcd.com/${getRandomInteger(0, 1000)}/info.0.json
-    const info: XkcdInfo = yield call(fetchAsJson, `https://xkcd.now.sh/?comic=${getRandomInteger(0, 1000)}`)
+    const info: XkcdInfo = yield call(fetchAsJson, `https://xkcd.vercel.app/?comic=${getRandomInteger(0, 1000)}`)
 
     yield put(XkcdReducerAction.setInfo(info))
   } catch (error: any) {
