@@ -1,4 +1,5 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+/* eslint-disable @typescript-eslint/no-var-requires */
+
 const withBundleAnalyzer = require('@next/bundle-analyzer')({
   enabled: process.env.ANALYZE === 'true',
 })
@@ -7,15 +8,7 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
  * @type {import('next').NextConfig}
  */
 const nextConfig = {
-  images: {
-    domains: ['pics.avs.io'],
-  },
   reactStrictMode: true,
-  // TODO: WIP
-  experimental: {
-    swcFileReading: false,
-    esmExternals: false,
-  },
 }
 
 module.exports = withBundleAnalyzer(nextConfig)
