@@ -1,4 +1,4 @@
-import { call, put, select, takeLatest } from 'typed-redux-saga'
+import { call, put, select, takeLatest } from 'redux-saga/effects'
 
 import type { DogResponse } from '@/eggs/dog/contracts/api-response'
 import { DogLoadingState } from '@/eggs/dog/contracts/state'
@@ -7,33 +7,30 @@ import { DogPublicAction, DogReducerAction } from '@/eggs/dog/slice'
 import { fetchAsJson } from '@/utils/fetch-as-json'
 
 function* loadDogWorker() {
-  yield* put(DogReducerAction.setLoadingState(DogLoadingState.LOADING))
+  yield put(DogReducerAction.setLoadingState(DogLoadingState.LOADING))
 
-  const error = yield* select(errorSelector)
+  const error: ReturnType<typeof errorSelector> = yield select(errorSelector)
 
   if (error) {
-    yield* put(DogReducerAction.setError(undefined))
+    yield put(DogReducerAction.setError(undefined))
   }
 
   try {
-    const res = yield* call<(...args: Parameters<typeof fetchAsJson>) => Promise<DogResponse>>(
-      fetchAsJson,
-      'https://dog.ceo/api/breeds/image/random',
-    )
+    const res: DogResponse = yield call(fetchAsJson, 'https://dog.ceo/api/breeds/image/random')
 
     if (res.status === 'success') {
-      yield* put(DogReducerAction.setDog(res.message))
+      yield put(DogReducerAction.setDog(res.message))
     } else {
-      yield* put(DogReducerAction.setError(res.message))
+      yield put(DogReducerAction.setError(res.message))
     }
   } catch (error: any) {
     console.error('[Error in `loadDogWorker`]', error)
-    yield* put(DogReducerAction.setError(error.message))
+    yield put(DogReducerAction.setError(error.message))
   } finally {
-    yield* put(DogReducerAction.setLoadingState(DogLoadingState.LOADED))
+    yield put(DogReducerAction.setLoadingState(DogLoadingState.LOADED))
   }
 }
 
 export function* loadDogWatcher() {
-  yield* takeLatest(DogPublicAction.loadDog, loadDogWorker)
+  yield takeLatest(DogPublicAction.loadDog, loadDogWorker)
 }
