@@ -1,4 +1,4 @@
-import { call, put, select, takeLatest } from 'typed-redux-saga'
+import { call, put, select, takeLatest } from 'redux-saga/effects'
 
 import type { PostsResponseItem } from '@/eggs/posts/contracts/api-response'
 import { PostsLoadingState } from '@/eggs/posts/contracts/state'
@@ -7,21 +7,18 @@ import { PostsPublicAction, PostsReducerAction } from '@/eggs/posts/slice'
 import { fetchAsJson } from '@/utils/fetch-as-json'
 
 function* loadPostsWorker() {
-  yield* put(PostsReducerAction.setLoadingState(PostsLoadingState.LOADING))
+  yield put(PostsReducerAction.setLoadingState(PostsLoadingState.LOADING))
 
-  const error = yield* select(errorSelector)
+  const error: ReturnType<typeof errorSelector> = yield select(errorSelector)
 
   if (error) {
-    yield* put(PostsReducerAction.setError(undefined))
+    yield put(PostsReducerAction.setError(undefined))
   }
 
   try {
-    const responseItems = yield* call<(...args: Parameters<typeof fetchAsJson>) => Promise<PostsResponseItem[]>>(
-      fetchAsJson,
-      'https://jsonplaceholder.typicode.com/posts',
-    )
+    const responseItems: PostsResponseItem[] = yield call(fetchAsJson, 'https://jsonplaceholder.typicode.com/posts')
 
-    yield* put(
+    yield put(
       PostsReducerAction.setPosts(
         responseItems.slice(0, 10).map(responseItem => ({
           id: `${responseItem.id}`,
@@ -31,12 +28,12 @@ function* loadPostsWorker() {
     )
   } catch (error: any) {
     console.error('[Error in `loadPostsWorker`]', error)
-    yield* put(PostsReducerAction.setError(error.message))
+    yield put(PostsReducerAction.setError(error.message))
   } finally {
-    yield* put(PostsReducerAction.setLoadingState(PostsLoadingState.LOADED))
+    yield put(PostsReducerAction.setLoadingState(PostsLoadingState.LOADED))
   }
 }
 
 export function* loadPostsWatcher() {
-  yield* takeLatest(PostsPublicAction.loadPosts, loadPostsWorker)
+  yield takeLatest(PostsPublicAction.loadPosts, loadPostsWorker)
 }
