@@ -1,4 +1,4 @@
-import { call, put, select, takeLatest } from 'typed-redux-saga'
+import { call, put, select, takeLatest } from 'redux-saga/effects'
 
 import type { XkcdInfo } from '@/eggs/xkcd/contracts/api-response'
 import { XkcdLoadingState } from '@/eggs/xkcd/contracts/state'
@@ -8,36 +8,33 @@ import { getRandomInteger } from '@/eggs/xkcd/utils/random-integer'
 import { fetchAsJson } from '@/utils/fetch-as-json'
 
 function* loadXkcdInfoWorker() {
-  yield* put(XkcdReducerAction.setLoadingState(XkcdLoadingState.LOADING))
+  yield put(XkcdReducerAction.setLoadingState(XkcdLoadingState.LOADING))
 
-  const currentInfo = yield* select(xkcdInfoSelector)
+  const currentInfo: ReturnType<typeof xkcdInfoSelector> = yield select(xkcdInfoSelector)
 
   if (currentInfo) {
-    yield* put(XkcdReducerAction.setInfo(undefined))
+    yield put(XkcdReducerAction.setInfo(undefined))
   }
 
-  const error = yield* select(errorSelector)
+  const error: ReturnType<typeof errorSelector> = yield select(errorSelector)
 
   if (error) {
-    yield* put(XkcdReducerAction.setError(undefined))
+    yield put(XkcdReducerAction.setError(undefined))
   }
 
   try {
     // Alternative api https://xkcd.com/${getRandomInteger(0, 1000)}/info.0.json
-    const info = yield* call<(...args: Parameters<typeof fetchAsJson>) => Promise<XkcdInfo>>(
-      fetchAsJson,
-      `https://xkcd.now.sh/?comic=${getRandomInteger(0, 1000)}`,
-    )
+    const info: XkcdInfo = yield call(fetchAsJson, `https://xkcd.vercel.app/?comic=${getRandomInteger(0, 1000)}`)
 
-    yield* put(XkcdReducerAction.setInfo(info))
+    yield put(XkcdReducerAction.setInfo(info))
   } catch (error: any) {
     console.error('[Error in `loadXkcdInfoWorker`]', error)
-    yield* put(XkcdReducerAction.setError(error.message))
+    yield put(XkcdReducerAction.setError(error.message))
   } finally {
-    yield* put(XkcdReducerAction.setLoadingState(XkcdLoadingState.LOADED))
+    yield put(XkcdReducerAction.setLoadingState(XkcdLoadingState.LOADED))
   }
 }
 
 export function* loadXkcdInfoWatcher() {
-  yield* takeLatest(XkcdPublicAction.loadInfo, loadXkcdInfoWorker)
+  yield takeLatest(XkcdPublicAction.loadInfo, loadXkcdInfoWorker)
 }
