@@ -1,8 +1,7 @@
 import type { EggTuple } from '@redux-eggs/core'
 
-import type { AppStore } from '../../index'
 import { getXkcdEgg } from '../xkcd'
 
-export const getCommonEggs = (): EggTuple<AppStore> => {
+export const getCommonEggs = (): EggTuple => {
   return [getXkcdEgg()]
 }
