@@ -1,7 +1,9 @@
-import type { Extension } from '@redux-eggs/core/dist'
+import type { Extension } from '@redux-eggs/core'
 import { createLogger } from 'redux-logger'
 
-export const getLoggerExtension = (): Extension<any> => {
+import type { AppStore } from './index'
+
+export const getLoggerExtension = (): Extension => {
   const logger = createLogger({
     level: {
       action() {
@@ -14,46 +16,38 @@ export const getLoggerExtension = (): Extension<any> => {
   })
 
   return {
-    middlewares: [logger],
+    middleware: logger,
 
-    beforeAdd: [
-      (eggs, store) => {
-        console.log(
-          'beforeAdd',
-          eggs.map(egg => egg.id),
-          store.getEggs().map(egg => egg.value.id),
-        )
-      },
-    ],
+    beforeAdd(eggs, store: AppStore) {
+      console.log(
+        'beforeAdd',
+        eggs.map(egg => egg.id),
+        store.getEggs().map(egg => egg.value.id),
+      )
+    },
 
-    afterAdd: [
-      (eggs, store) => {
-        console.log(
-          'afterAdd',
-          eggs.map(egg => egg.id),
-          store.getEggs().map(egg => egg.value.id),
-        )
-      },
-    ],
+    afterAdd(eggs, store: AppStore) {
+      console.log(
+        'afterAdd',
+        eggs.map(egg => egg.id),
+        store.getEggs().map(egg => egg.value.id),
+      )
+    },
 
-    beforeRemove: [
-      (eggs, store) => {
-        console.log(
-          'beforeRemove',
-          eggs.map(egg => egg.id),
-          store.getEggs().map(egg => egg.value.id),
-        )
-      },
-    ],
+    beforeRemove(eggs, store: AppStore) {
+      console.log(
+        'beforeRemove',
+        eggs.map(egg => egg.id),
+        store.getEggs().map(egg => egg.value.id),
+      )
+    },
 
-    afterRemove: [
-      (eggs, store) => {
-        console.log(
-          'afterRemove',
-          eggs.map(egg => egg.id),
-          store.getEggs().map(egg => egg.value.id),
-        )
-      },
-    ],
+    afterRemove(eggs, store: AppStore) {
+      console.log(
+        'afterRemove',
+        eggs.map(egg => egg.id),
+        store.getEggs().map(egg => egg.value.id),
+      )
+    },
   }
 }
