@@ -1,192 +1,153 @@
-import type { Reducer, ReducersMapObject } from 'redux'
-import { combineReducers, createStore } from 'redux'
-
-import type { ReducerEntry } from '@/reducer-tray'
+import type { Counter, CounterItem, ReducerEntry } from '@/contracts'
+import * as CounterModule from '@/counter'
 import { getReducerTray } from '@/reducer-tray'
-import { REDUCE_ACTION_TYPE } from '@/store'
 
 describe('Tests for reducer tray', () => {
-  const getReducer = (): Reducer => {
-    return (state = { test: true }) => {
-      return state
-    }
+  const buildCounterGetter = () => {
+    let mockedCounter: Counter<string>
+
+    const { getCounter } = jest.requireActual('@/counter')
+
+    jest.spyOn(CounterModule, 'getCounter').mockImplementationOnce((...args) => {
+      mockedCounter = getCounter(...args)
+      return mockedCounter
+    })
+
+    return () => mockedCounter
   }
 
-  const entry1: ReducerEntry = ['reducer1', getReducer()]
-  const entry2: ReducerEntry = ['reducer2', getReducer()]
-  const entry3: ReducerEntry = ['reducer3', getReducer()]
-  const entry4: ReducerEntry = ['reducer4', getReducer()]
-
-  const spyOnConsoleWarn = jest.spyOn(console, 'warn')
-  const spyOnConsoleError = jest.spyOn(console, 'error')
-
-  afterEach(() => {
-    expect(spyOnConsoleWarn).not.toBeCalled()
-    expect(spyOnConsoleError).not.toBeCalled()
-  })
-
-  afterAll(() => {
-    spyOnConsoleWarn.mockRestore()
-    spyOnConsoleError.mockRestore()
-  })
-
   test('Reducers should be added to and removed from tray correctly', () => {
-    const tray = getReducerTray(combineReducers)
+    const entry1: ReducerEntry = ['reducer1', () => ({})]
+    const entry2: ReducerEntry = ['reducer2', () => ({})]
 
-    const store = createStore(tray.reducer)
+    const getCounter = buildCounterGetter()
 
-    tray.add([entry1])
-    tray.add([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer1', 'reducer2'])
+    const tray = getReducerTray(() => () => ({}))
+    expect(getCounter().getItems()).toEqual([])
 
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer1'])
+    tray.add([])
+    expect(getCounter().getItems()).toEqual([])
 
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer1'])
+    tray.remove([])
+    expect(getCounter().getItems()).toEqual([])
 
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer1'])
+    tray.add([entry1, entry2])
+    expect(getCounter().getCount(entry1[0])).toEqual(1)
+    expect(getCounter().getCount(entry2[0])).toEqual(1)
 
-    tray.add([entry3])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer1', 'reducer3'])
+    tray.add([entry1, entry2])
+    expect(getCounter().getCount(entry1[0])).toEqual(2)
+    expect(getCounter().getCount(entry2[0])).toEqual(2)
 
-    // Reducer count: `0`
-    tray.remove([entry1])
-    tray.remove([entry3])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual([])
+    const fullTray: CounterItem<string>[] = [
+      { count: 2, value: entry1[0] },
+      { count: 2, value: entry2[0] },
+    ]
 
-    // Reducer count: `0`
-    tray.remove([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual([])
+    tray.add([])
+    expect(getCounter().getItems()).toEqual(fullTray)
 
-    // Reducer count: `1`
-    tray.add([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer4'])
+    tray.remove([])
+    expect(getCounter().getItems()).toEqual(fullTray)
 
-    // Reducer count: `2`
-    tray.add([entry4])
-    // Reducer count: `3`
-    tray.add([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer4'])
+    tray.remove([entry1, entry2])
+    expect(getCounter().getCount(entry1[0])).toEqual(1)
+    expect(getCounter().getCount(entry2[0])).toEqual(1)
 
-    // Reducer count: `2`
-    tray.remove([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer4'])
+    tray.remove([entry1, entry2])
+    expect(getCounter().getCount(entry1[0])).toEqual(0)
+    expect(getCounter().getCount(entry2[0])).toEqual(0)
 
-    // Reducer count: `1`
-    tray.remove([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual(['reducer4'])
+    tray.remove([entry1, entry2])
+    expect(getCounter().getItems()).toEqual([])
 
-    // Reducer count: `0`
-    tray.remove([entry4])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(Object.keys(store.getState())).toEqual([])
+    tray.add([entry1, entry2, entry1, entry2])
+    expect(getCounter().getCount(entry1[0])).toEqual(2)
+    expect(getCounter().getCount(entry2[0])).toEqual(2)
+
+    tray.remove([entry1, entry2, entry1, entry2])
+    expect(getCounter().getItems()).toEqual([])
   })
 
   test('Reducer tray works correctly with custom reducer combiner', () => {
-    let combinedKeys: string[] = []
+    const anyState = {}
+    const anyAction = { type: 'action' } as const
 
+    const entry1: ReducerEntry = ['reducer1', jest.fn()]
+    const entry2: ReducerEntry = ['reducer2', jest.fn()]
+
+    const combinerResult = jest.fn()
     const combiner = jest.fn()
-    combiner.mockImplementation((reducersMap: ReducersMapObject): ReturnType<typeof combineReducers> => {
-      combinedKeys = Object.keys(reducersMap)
-      return combineReducers(reducersMap)
-    })
+    combiner.mockReturnValue(combinerResult)
 
     const tray = getReducerTray(combiner)
 
-    const store = createStore(tray.reducer)
-
+    tray.add([])
+    tray.dynamicReducer(anyState, anyAction)
     expect(combiner).not.toBeCalled()
-    expect(combinedKeys).toEqual([])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
 
-    tray.add([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.add([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(combiner).toBeCalledTimes(2)
-    expect(combinedKeys).toEqual(['reducer1', 'reducer2'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    tray.remove([])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).not.toBeCalled()
 
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
+    tray.add([entry1, entry2])
+    tray.dynamicReducer(undefined, anyAction)
     expect(combiner).toBeCalledTimes(1)
-    expect(combinedKeys).toEqual(['reducer1'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
-
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(combiner).not.toBeCalled()
-    expect(combinedKeys).toEqual(['reducer1'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
-
-    tray.remove([entry3])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(combiner).not.toBeCalled()
-    expect(combinedKeys).toEqual(['reducer1'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
-
-    tray.add([entry3])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
+    expect(combiner).nthCalledWith(1, { [entry1[0]]: entry1[1], [entry2[0]]: entry2[1] })
+    expect(combinerResult).toBeCalledTimes(1)
+    expect(combinerResult).nthCalledWith(1, anyState, anyAction)
+    tray.dynamicReducer(anyState, anyAction)
     expect(combiner).toBeCalledTimes(1)
-    expect(combinedKeys).toEqual(['reducer1', 'reducer3'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    expect(combiner).nthCalledWith(1, { [entry1[0]]: entry1[1], [entry2[0]]: entry2[1] })
+    expect(combinerResult).toBeCalledTimes(2)
+    expect(combinerResult).nthCalledWith(2, anyState, anyAction)
 
-    tray.remove([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.remove([entry3])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
+    tray.add([entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(1)
+    expect(combinerResult).toBeCalledTimes(3)
+    expect(combinerResult).nthCalledWith(3, anyState, anyAction)
+
+    tray.add([])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(1)
+    expect(combinerResult).toBeCalledTimes(4)
+    expect(combinerResult).nthCalledWith(4, anyState, anyAction)
+
+    tray.remove([])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(1)
+    expect(combinerResult).toBeCalledTimes(5)
+    expect(combinerResult).nthCalledWith(5, anyState, anyAction)
+
+    tray.remove([entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(1)
+    expect(combinerResult).toBeCalledTimes(6)
+    expect(combinerResult).nthCalledWith(6, anyState, anyAction)
+
+    tray.remove([entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
     expect(combiner).toBeCalledTimes(2)
-    expect(combinedKeys).toEqual([])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    expect(combiner).nthCalledWith(2, {})
+    expect(combinerResult).toBeCalledTimes(6)
 
-    tray.add([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.add([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.add([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.add([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
+    tray.remove([entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
     expect(combiner).toBeCalledTimes(2)
-    expect(combinedKeys).toEqual(['reducer1', 'reducer2'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    expect(combinerResult).toBeCalledTimes(6)
 
-    tray.remove([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(combiner).not.toBeCalled()
-    expect(combinedKeys).toEqual(['reducer1', 'reducer2'])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    tray.add([entry1, entry2, entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(3)
+    expect(combiner).nthCalledWith(3, { [entry1[0]]: entry1[1], [entry2[0]]: entry2[1] })
+    expect(combinerResult).toBeCalledTimes(7)
+    expect(combinerResult).nthCalledWith(7, anyState, anyAction)
 
-    tray.remove([entry1])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    tray.remove([entry2])
-    store.dispatch({ type: REDUCE_ACTION_TYPE })
-    expect(combiner).toBeCalledTimes(2)
-    expect(combinedKeys).toEqual([])
-    expect(Object.keys(store.getState())).toEqual(combinedKeys)
-    combiner.mockClear()
+    tray.remove([entry1, entry2, entry1, entry2])
+    tray.dynamicReducer(anyState, anyAction)
+    expect(combiner).toBeCalledTimes(4)
+    expect(combiner).nthCalledWith(4, {})
+    expect(combinerResult).toBeCalledTimes(7)
   })
 })
