@@ -1,2 +1,3 @@
-export * from '@/injector'
-export * from '@/with-eggs'
+export type { InjectorResult, WithEggsReturnType } from '@/contracts'
+export { getInjector } from '@/injector'
+export { withEggs } from '@/with-eggs'
