@@ -1,4 +1,15 @@
-import type { Saga, Task } from 'redux-saga'
+import type { Extension } from '@redux-eggs/core'
+import type { StoreEnhancer } from 'redux'
+import type { Saga, SagaMiddleware, Task } from 'redux-saga'
+
+// TODO: Pass arguments with Saga's
+export interface SagaTray {
+  getTasks(): Task[]
+  add(sagas: Saga[]): void
+  remove(sagas: Saga[]): void
+}
+
+export type SagaExtension = Extension<SagaMiddleware, StoreEnhancer<SagaExt>>
 
 export interface SagaExt {
   getSagaTasks: () => Task[]
@@ -7,9 +18,5 @@ export interface SagaExt {
 declare module '@redux-eggs/core' {
   interface Egg {
     sagas?: Saga[]
-  }
-
-  interface EggExt {
-    getSagaTasks: () => Task[]
   }
 }
