@@ -1,20 +1,12 @@
-import type { Egg, Extension } from '@redux-eggs/core'
-import type { ExtensionEventHandler } from '@redux-eggs/core'
+import type { Egg } from '@redux-eggs/core'
 import type { StoreEnhancer } from 'redux'
-import type { Saga, SagaMiddleware, SagaMiddlewareOptions } from 'redux-saga'
+import type { Saga, SagaMiddlewareOptions } from 'redux-saga'
 import createSagaMiddleware from 'redux-saga'
 
-import type { SagaExt } from '@/contracts'
+import type { SagaExt, SagaExtension } from '@/contracts'
 import { getSagaTray } from '@/tray'
 
-export interface SagaExtension extends Extension<any> {
-  middlewares: SagaMiddleware[]
-  enhancers: StoreEnhancer<SagaExt>[]
-  afterAdd: ExtensionEventHandler<any>[]
-  afterRemove: ExtensionEventHandler<any>[]
-}
-
-const getSagas = (eggs: Egg<any>[]) => {
+const getSagas = (eggs: Egg[]): Saga[] => {
   return eggs.reduce<Saga[]>((acc, egg) => {
     if (egg.sagas?.length) {
       acc.push(...egg.sagas)
@@ -23,22 +15,24 @@ const getSagas = (eggs: Egg<any>[]) => {
   }, [])
 }
 
-export const getSagaExtension = (options?: SagaMiddlewareOptions): SagaExtension => {
-  const middleware: SagaMiddleware = createSagaMiddleware(options)
+export const getSagaExtension = <C extends object = {}>(options?: SagaMiddlewareOptions<C>): SagaExtension => {
+  const middleware = createSagaMiddleware<C>(options)
 
   const tray = getSagaTray(middleware)
 
-  const enhancer: StoreEnhancer<SagaExt> = createStore => (reducer, preloadedState) => {
-    return Object.assign({}, createStore(reducer, preloadedState), { getSagaTasks: tray.getTasks })
-  }
+  const enhancer: StoreEnhancer<SagaExt> = createStore => (reducer, preloadedState) => ({
+    ...createStore(reducer, preloadedState),
+    getSagaTasks: tray.getTasks,
+  })
 
   return {
-    middlewares: [middleware],
-
-    enhancers: [enhancer],
-
-    afterAdd: [(eggs: Egg<any>[]): void => tray.add(getSagas(eggs))],
-
-    afterRemove: [(eggs: Egg<any>[]): void => tray.remove(getSagas(eggs))],
+    middleware,
+    enhancer,
+    afterAdd(eggs) {
+      tray.add(getSagas(eggs))
+    },
+    afterRemove(eggs) {
+      tray.remove(getSagas(eggs))
+    },
   }
 }
